DEA

Performs differential expression analysis for proteomes (and modified proteomes) using limma

Usage

DEA(
  prot.Data = NULL,
  enrich.Data = NULL,
  sampleTable,
  fasta = NULL,
  org = "dme",
  quantification = "LFQ",
  pvalCutOff = 0.05,
  sigmaCutOff = 0.05,
  lfcCutOff = 0,
  contrasts,
  Fraction = c("Proteome", "Enriched"),
  filter.protein.type = "condition",
  filter.thr = 0,
  filter.protein.min = 0.75,
  probability = NULL,
  enrich.batch = FALSE,
  impute.function = c("knn", "MLE", "QRILC", "man", "MinProb", "bpca", "MinDet"),
  q.MinProbDet = 0.01,
  k.knn = 10,
  rowmax.knn = 0.9,
  shift.man = 1.8,
  scale.man = 0.3,
  distance.matrix = "spearman",
  clustering.method = "ward.D2",
  fav.proteins = NULL,
  name.sigProteins = FALSE,
  timeSeries = FALSE,
  enrich = "gsea",
  rankBy = "stat",
  KEGG = FALSE,
  ont = "BP",
  padjustMethod.enrich = "fdr",
  background = TRUE,
  minGS = 50,
  maxGS = 500,
  simplify = FALSE,
  simplify_cutoff = 0.7,
  plotType = c("dotPlot", "cNetPlot", "heatPlot", "treePlot", "gseaPlot", "ridgePlot"),
  circular = FALSE,
  colorEdge = FALSE,
  nodeLabel = c("gene", "category", "all", "none"),
  cexLabelCategory = 1.2,
  cexLabelGene = 0.8,
  colorCcategory = "black",
  colorGene = "black",
  showCategory = 10,
  aa = "K",
  seq.width = 15,
  min.seqs = 5,
  motif.pval = 1e-05
)

Arguments

prot.Data

Input proteome output from MaxQuant

enrich.Data

Modified proteome output from MaxQuant. Default is NULL

sampleTable

.xlsx file containing information about the samples. Three columns are mandatory (label, condition and replicate)

fasta

fasta file from uniprot (same fasta flies used for MaxQuant search). Default is NULL. Fasta file required for motif search

org

Database of the organism. Drosophila melanogaster = "dme", Mus muscuslus ' "mmu", Homo sapiens = "hsa", Saccharomyces cerevisae = "sce". Default is "dme"

quantification

Default is LFQ. Can be either "LFQ" or "iBAQ"

pvalCutOff

P-value cut off for significant protein(s), site(s) and GO terms. Default is 0.05

sigmaCutOff

PI-value cut off for significant protein(s) and site(s). Default is 0.05 (Refer Xiao et al, 2014 and Hostrup et al, 2022 for details on PI-value)

lfcCutOff

Log-fold change cut off. Default is 0

contrasts

Mentions the conditions to be compared. Ex. MUTANT_vs_WILDTYPE or (MUTANT-A_vs_WILDTYPE-A)_vs_(MUTANT-B_vs_WILDTYPE-B). This how the contrasts should be provided and the values should be the same as the one given in condition column of sampleTable. (MUTANT-A_vs_WILDTYPE-A)_vs_(MUTANT-B_vs_WILDTYPE-B): This type can be used for complex data when comparing the interaction between two conditions such genotype and time

Fraction

Can be either "Proteome" or "Enriched". Indicates the type of input data used

filter.protein.type

Can be either "complete" or "condition" or "fraction". complete indictaes removal of all NAs. condition indicates removal of NAs based on different conditions in the data (Ex. mutant and control). fraction indicates removal of NAs based on all samples irrespective of different conditions (check DEP package for further details)

filter.thr

Only if filter.protein.type = condition. Numerical value less than the number of relicates in either condition (Ex. 0 indicates the protein should have no NAs in all replicates of atleast one condition while 1 indicates they can have one NAs)

filter.protein.min

Only if filter.protein.type = fraction. Any value between 0-1 Any value between 0-1 (Ex. 0.75 indicates the protein should not have NAs in 75% of all samples)

probability

Numeric value between 0-1. Filters out modified peptides with probabilities less than the given value (Only used if Fraction = "Enriched")

enrich.batch

If the normalization should be done based on paired replicates or avergae of replicates

impute.function

If filter.protein.type = "condition", one of "QRILC", "man", "MinProb", "MinDet". If filter.protein.type = "fraction", one of "bpca", "knn", "MLE"

q.MinProbDet

Default is 0.01 (Refer DEP package for further information)

k.knn

Default is 10 (Refer DEP package for further information)

rowmax.knn

Default in 0.9 (Refer DEP package for further information)

shift.man

Default is 1.8 (Refer DEP package for further information)

scale.man

Deafult is 0.3 (Refer DEP package for further information)

distance.matrix

One of "spearman", "pearson", "uncentered correlation", "absolute pearson", "sqrt", "weird". Used for heatmap and default is spearman

clustering.method

One of "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid". Used for heatmap and default is ward.D2

fav.proteins

Vector with favourite proteins (or their corresponding sites) to be named in the volcano plot irrespective of its statistical significance. Also used for plotting corresponding bar plot/timeseries plot (in that case, only significant ones will be plotted)

name.sigProteins

Logical (Default is FALSE). If TRUE, names all significant protein(s) or site(s)

timeSeries

Logical (Default is FALSE). If TRUE, line plots will be generated instead of barplots

enrich

Can be either "gsea" or "ora". Default is gsea

rankBy

Can be either stat or logFC (stat indicates t.statistic value). Genes are ranked accordingly for GSEA. Default is stat

KEGG

Logical, default is FALSE

ont

Can be "BP", "CC", "MF" or "ALL"

padjustMethod.enrich

Method for adjusting p-value (Refer clusterProfiler for available options). Default is fdr

background

Logical, default is TRUE. If FALSE, background proteins will not be considered for over representation analysis

minGS

minimal size of annotated genes in geneset

maxGS

maximal size of annotated genes in geneset

simplify

Logical, default is FALSE. If TRUE, redundant GO terms will be simplified based on the simplify_cutoff

simplify_cutoff

Cutoff for simplify method. Numerical value from 0-1. Default is 0.7

plotType

Can be one of "dotPlot", "cNetPlot", "heatPlot", "treePlot", "gseaPlot", "ridgePlot"

circular

Logical, default is FALSE. If TRUE, circular visualization will be used for cNetPlot

colorEdge

Logical, default is FALSE. Parameter to be used if circular = "TRUE" and plotType ' "cNetPlot

nodeLabel

Can be one of "gene", "category", "all", "none". Parameter to be used if circular = "TRUE" and plotType ' "cNetPlot

cexLabelCategory

Parameter to be used if circular = "TRUE" and plotType ' "cNetPlot (Check enrichplot package for further info). Defaults i 1.2

cexLabelGene

Parameter to be used if circular = "TRUE" and plotType ' "cNetPlot (Check enrichplot package for further info). Default is 0.8

colorCcategory

Parameter to be used if circular = "TRUE" and plotType ' "cNetPlot (Check enrichplot package for further info). Default is "black"

colorGene

Parameter to be used if circular = "TRUE" and plotType ' "cNetPlot (Check enrichplot package for further info). Default is "black"

showCategory

Number of GO terms to be displayed in the plot. Default is 10

aa

Central amino acid. For example, "K" (for acetylome), "S", "T", "Y", "STY" (for phosphoproteome). Default is "K".

seq.width

Width of the sequence for motif search. Default is 15

min.seqs

This threshold refers to the minimum number of times you wish each of your extracted motifs to occur in the data set. Default is 20 and is usually appropriate, although this parameter may be adjusted to yield more specific or less specific motifs.

motif.pval

The p-value threshold for the binomial probability. This is used for the selection of significant residue/position pairs in the motif. Default is 1e-05 and is suggested to maintain a low false positive rate in standard protein motif analyses.

Value

Returns a complete comprenhensive analysis separated into different folders for each type of analysis from quality control, filtering, limma differential expression analysis, GO term analysis and visualization and motif analysis for modified proteomes