EnrichmentAnalysis

Performs Gene Set Enrichment Analysis and Over Representation Analaysis using clusterProfiler package

EnrichmentAnalysis(
  enrich = c("gsea", "ora"),
  nonExclusive.data = NULL,
  Fraction,
  rankBy = "stat",
  KEGG = FALSE,
  ont = "BP",
  padjustMethod = "fdr",
  exclusive.data = NULL,
  background = TRUE,
  minGS = 50,
  maxGS = 500,
  simplify = FALSE,
  simplify_cutoff = 0.7,
  org = "dme",
  contrasts,
  pvalCutOff = 0.05,
  sigmaCutOff = 0.05,
  lfcCutOff = 0
)

Arguments

enrich

Can be either "gsea" or "enrich"

nonExclusive.data

Default is NULL. If not, the data should be given as list containing data frame(s)

Fraction

Can be either "Proteome" or "Enriched". Indicates the type of input data used

rankBy

Can be either stat or logFC. Genes are ranked accordingly for GSEA

KEGG

Logical, default is FALSE

ont

Can be "BP", "CC", "MF" or "ALL"

padjustMethod

Method for adjusting p-value (Refer clusterProfiler for available options)

exclusive.data

Default is NULL. If not, the data should be given as list containing data frame(s)

background

Logical, default is TRUE. If FALSE, background proteins will not be considered for over representation analysis

minGS

minimal size of annotated genes in geneset

maxGS

maximal size of annotated genes in geneset

simplify

Logical, default is FALSE. If TRUE, redundant GO terms will be simplified based on the simplify_cutoff

simplify_cutoff

Cutoff for simplify method. Numerical value from 0-1. Default is 0.7

org

Database of the organism. Drosophila melanogaster = "dme", Mus muscuslus ' "mmu", Homo sapiens = "hsa", Saccharomyces cerevisae = "sce". Default is "dme"

contrasts

Mentions the conditions to be compared. Ex. MUTANT_vs_WILDTYPE or (MUTANT-A_vs_WILDTYPE-A)_vs_(MUTANT-B_vs_WILDTYPE-B). This how the contrasts should be provided and the values should be the same as the one given in condition column of sampleTable. (MUTANT-A_vs_WILDTYPE-A)_vs_(MUTANT-B_vs_WILDTYPE-B): This type can be used for complex data when comparing the interaction between two conditions such genotype and time

pvalCutOff

P-value cut off for significant protein(s), site(s) and GO terms. Default is 0.05

sigmaCutOff

PI-value cut off for significant protein(s) and site(s). Default is 0.05 (Refer Xiao et al, 2014 and Hostrup et al, 2022 for details on PI-value)

lfcCutOff

Log-fold change cut off. Default is 0

Value

Generates list of significant GO terms