editData

Cleans the raw data from MaxQuant for downstream analysis

Usage

editData(data, Fraction, probability, org, quantification)

Arguments

data

Input or modified proteome file from MaxQuant (.txt)

Fraction

Can be either "Proteome" or "Enriched". Indicates the type of input data used

probability

Numeric value between 0-1. Filters out modified peptides with probabilities less than the given value (Only used if Fraction = "Enriched")

org

Database of the organism. Drosophila melanogaster = "dme", Mus muscuslus ' "mmu", Homo sapiens = "hsa", Saccharomyces cerevisae = "sce". Default is "dme"

quantification

Default is LFQ. Can be either "LFQ" or "iBAQ"

Value

A data frame that will be cleaner than the original raw data to be used for differential expression analysis

Examples

lfqdata <- editData(data = proteome, Fraction = "Enriched", probability = 0.9)
#> Error in editData(data = proteome, Fraction = "Enriched", probability = 0.9): argument "org" is missing, with no default