Modified proteome data

This file is obtained from MaxQuant as Acetyl(K)Sites.txt. Column information is not explained as these can be obtained in the summary.pdf file from MaxQuant output This corresponding acetylome data is obtained from a recently published study comparing the whole cell extracts of acetylomes between wild-type and chameau RNAi flies (Further details can be found in the link provided below)

Usage

data(enrich.Data)

Format

An object of class "data.frame"

Source

EMBO Reports

References

Venkatasubramani AV, et al., (2023), EMBO Rep Oct 9;24(10):e57023

Examples

data(enrich.Data)
#> Warning: data set ‘enrich.Data’ not found
head(enrich.Data)
#>                                                                                                                                                                                                                                                        Proteins
#> 1                                                                                                                                                                                                                                                   CON__P02662
#> 2                                                                                                                                                                                                                                                   CON__P02663
#> 3                                                                                                                                                                                                                                                   CON__P02666
#> 4 CON__P04259;CON__P48668;CON__P02538;CON__P07744;CON__Q3KNV1;CON__P08729;CON__P12035;CON__P35908v2;CON__P35908;CON__O95678;CON__Q01546;CON__Q3TTY5;CON__Q5XKE5;CON__Q5XQN5;CON__Q922U2;CON__P13647;CON__Q8VED5;CON__Q8BGZ7;CON__Q6IFZ6;CON__Q9NSB2;CON__Q6ISB0
#> 5                                                                                                                                                                                                                                                   CON__P06868
#> 6                                                                                                                                                                                                                                                              
#>                                                             Positions.within.proteins
#> 1                                                                                  83
#> 2                                                                                 181
#> 3                                                                                 176
#> 4 173;173;173;156;101;101;208;188;194;159;193;209;152;179;172;178;149;160;177;175;175
#> 5                                                                                   7
#> 6                                                                                    
#>                                                                                                            Leading.proteins
#> 1                                                                                                               CON__P02662
#> 2                                                                                                               CON__P02663
#> 3                                                                                                               CON__P02666
#> 4 CON__P04259;CON__P07744;CON__Q3KNV1;CON__P12035;CON__P35908v2;CON__Q3TTY5;CON__Q5XKE5;CON__Q5XQN5;CON__Q6IFZ6;CON__Q9NSB2
#> 5                                                                                                               CON__P06868
#> 6                                                                                                 REV__sp|B7YZV4|PDE1_DROME
#>                     Protein Fasta.headers Localization.prob Score.diff
#> 1               CON__P02662                        1.000000    76.4649
#> 2               CON__P02663                        1.000000    82.9077
#> 3               CON__P02666                        1.000000    41.0916
#> 4             CON__P35908v2                        0.999984    48.0728
#> 5               CON__P06868                        1.000000    60.9103
#> 6 REV__sp|B7YZV4|PDE1_DROME                        1.000000   102.0610
#>           PEP   Score Delta.score Score.for.localization
#> 1 0.029754500  76.465      37.317                 76.465
#> 2 0.000698673  82.908      73.327                 82.908
#> 3 0.004190350  86.624      80.384                 41.092
#> 4 0.027041100  58.831      46.546                 58.831
#> 5 0.000493173  84.498      68.868                 60.910
#> 6 0.027045900 102.060      39.239                102.060
#>   Localization.prob.C1_IP Score.diff.C1_IP PEP.C1_IP Score.C1_IP
#> 1                      NA               NA        NA          NA
#> 2                      NA               NA        NA          NA
#> 3                      NA               NA        NA          NA
#> 4                      NA               NA        NA          NA
#> 5                      NA               NA        NA          NA
#> 6                      NA               NA        NA          NA
#>   Localization.prob.C1_prot Score.diff.C1_prot PEP.C1_prot Score.C1_prot
#> 1                        NA                 NA          NA            NA
#> 2                        NA                 NA          NA            NA
#> 3                        NA                 NA          NA            NA
#> 4                        NA                 NA          NA            NA
#> 5                         0                  0          NA           NaN
#> 6                        NA                 NA          NA            NA
#>   Localization.prob.C2_IP Score.diff.C2_IP PEP.C2_IP Score.C2_IP
#> 1                      NA               NA        NA          NA
#> 2                      NA               NA        NA          NA
#> 3                      NA               NA        NA          NA
#> 4                      NA               NA        NA          NA
#> 5                      NA               NA        NA          NA
#> 6                      NA               NA        NA          NA
#>   Localization.prob.C2_prot Score.diff.C2_prot PEP.C2_prot Score.C2_prot
#> 1                        NA                 NA          NA            NA
#> 2                        NA                 NA          NA            NA
#> 3                        NA                 NA          NA            NA
#> 4                        NA                 NA          NA            NA
#> 5                         0                  0          NA           NaN
#> 6                         0                  0          NA           NaN
#>   Localization.prob.C3_IP Score.diff.C3_IP PEP.C3_IP Score.C3_IP
#> 1                      NA               NA        NA          NA
#> 2                      NA               NA        NA          NA
#> 3                      NA               NA        NA          NA
#> 4                      NA               NA        NA          NA
#> 5                      NA               NA        NA          NA
#> 6                      NA               NA        NA          NA
#>   Localization.prob.C3_prot Score.diff.C3_prot PEP.C3_prot Score.C3_prot
#> 1                        NA                 NA          NA            NA
#> 2                        NA                 NA          NA            NA
#> 3                        NA                 NA          NA            NA
#> 4                        NA                 NA          NA            NA
#> 5                         1            66.3511  0.00128789        66.351
#> 6                        NA                 NA          NA            NA
#>   Localization.prob.C4_IP Score.diff.C4_IP PEP.C4_IP Score.C4_IP
#> 1                      NA               NA        NA          NA
#> 2                      NA               NA        NA          NA
#> 3                      NA               NA        NA          NA
#> 4                      NA               NA        NA          NA
#> 5                      NA               NA        NA          NA
#> 6                      NA               NA        NA          NA
#>   Localization.prob.C4_prot Score.diff.C4_prot PEP.C4_prot Score.C4_prot
#> 1                        NA                 NA          NA            NA
#> 2                        NA                 NA          NA            NA
#> 3                        NA                 NA          NA            NA
#> 4                        NA                 NA          NA            NA
#> 5                         1            84.4978 0.000493173        84.498
#> 6                        NA                 NA          NA            NA
#>   Localization.prob.C5_IP Score.diff.C5_IP PEP.C5_IP Score.C5_IP
#> 1                      NA               NA        NA          NA
#> 2                      NA               NA        NA          NA
#> 3                      NA               NA        NA          NA
#> 4                      NA               NA        NA          NA
#> 5                      NA               NA        NA          NA
#> 6                      NA               NA        NA          NA
#>   Localization.prob.C5_prot Score.diff.C5_prot PEP.C5_prot Score.C5_prot
#> 1                        NA                 NA          NA            NA
#> 2                        NA                 NA          NA            NA
#> 3                        NA                 NA          NA            NA
#> 4                        NA                 NA          NA            NA
#> 5                         1            52.4819  0.00457332        52.482
#> 6                        NA                 NA          NA            NA
#>   Localization.prob.W1_IP Score.diff.W1_IP PEP.W1_IP Score.W1_IP
#> 1                      NA               NA        NA          NA
#> 2                      NA               NA        NA          NA
#> 3                      NA               NA        NA          NA
#> 4                      NA               NA        NA          NA
#> 5                      NA               NA        NA          NA
#> 6                      NA               NA        NA          NA
#>   Localization.prob.W1_prot Score.diff.W1_prot PEP.W1_prot Score.W1_prot
#> 1                        NA                 NA          NA            NA
#> 2                        NA                 NA          NA            NA
#> 3                        NA                 NA          NA            NA
#> 4                        NA                 NA          NA            NA
#> 5                         1            60.9103  0.00194152         60.91
#> 6                        NA                 NA          NA            NA
#>   Localization.prob.W2_IP Score.diff.W2_IP PEP.W2_IP Score.W2_IP
#> 1                      NA               NA        NA          NA
#> 2                      NA               NA        NA          NA
#> 3                1.000000          59.9076 0.0507015      59.908
#> 4                0.999984          48.0728 0.0270411      58.831
#> 5                      NA               NA        NA          NA
#> 6                      NA               NA        NA          NA
#>   Localization.prob.W2_prot Score.diff.W2_prot PEP.W2_prot Score.W2_prot
#> 1                        NA                 NA          NA            NA
#> 2                        NA                 NA          NA            NA
#> 3                         0              0.000          NA           NaN
#> 4                        NA                 NA          NA            NA
#> 5                         1             52.255  0.00467139        52.255
#> 6                         0              0.000          NA           NaN
#>   Localization.prob.W3_IP Score.diff.W3_IP PEP.W3_IP Score.W3_IP
#> 1                      NA               NA        NA          NA
#> 2                      NA               NA        NA          NA
#> 3                      NA               NA        NA          NA
#> 4                      NA               NA        NA          NA
#> 5                      NA               NA        NA          NA
#> 6                      NA               NA        NA          NA
#>   Localization.prob.W3_prot Score.diff.W3_prot PEP.W3_prot Score.W3_prot
#> 1                        NA                 NA          NA            NA
#> 2                        NA                 NA          NA            NA
#> 3                        NA                 NA          NA            NA
#> 4                        NA                 NA          NA            NA
#> 5                         1            65.2463  0.00142061        65.246
#> 6                         1           102.0610  0.02704590       102.060
#>   Localization.prob.W4_IP Score.diff.W4_IP   PEP.W4_IP Score.W4_IP
#> 1                       1          76.4649 0.029754500      76.465
#> 2                       1          82.9077 0.000698673      82.908
#> 3                       1          41.0916 0.004190350      86.624
#> 4                      NA               NA          NA          NA
#> 5                      NA               NA          NA          NA
#> 6                      NA               NA          NA          NA
#>   Localization.prob.W4_prot Score.diff.W4_prot PEP.W4_prot Score.W4_prot
#> 1                        NA                 NA          NA            NA
#> 2                        NA                 NA          NA            NA
#> 3                        NA                 NA          NA            NA
#> 4                        NA                 NA          NA            NA
#> 5                         1            59.3497  0.00212899         59.35
#> 6                         0             0.0000          NA           NaN
#>   Localization.prob.W5_IP Score.diff.W5_IP PEP.W5_IP Score.W5_IP
#> 1                      NA               NA        NA          NA
#> 2                      NA               NA        NA          NA
#> 3                      NA               NA        NA          NA
#> 4                      NA               NA        NA          NA
#> 5                      NA               NA        NA          NA
#> 6                      NA               NA        NA          NA
#>   Localization.prob.W5_prot Score.diff.W5_prot PEP.W5_prot Score.W5_prot
#> 1                        NA                 NA          NA            NA
#> 2                        NA                 NA          NA            NA
#> 3                        NA                 NA          NA            NA
#> 4                        NA                 NA          NA            NA
#> 5                         1            60.9103  0.00194152         60.91
#> 6                         0             0.0000          NA           NaN
#>   Diagnostic.peak Number.of.Acetyl..K. Amino.acid
#> 1               +                    1          K
#> 2               +                    1          K
#> 3               +                    1          K
#> 4               +                    1          K
#> 5                                    1          K
#> 6               +                    1          K
#>                                                                                                                                                                                                                                                                                                                   Sequence.window
#> 1                                                                                                                                                                                                                                                                                                 SEEIVPNSVEQKHIQKEDVPSERYLGYLEQL
#> 2                                                                                                                                                                                                                                                                                                 KISQRYQKFALPQYLKTVYQHQKAMKPWIQP
#> 3                                                                                                                                                                                                                                                                                                 SVLSLSQSKVLPVPQKAVPYPQRDMPIQAFL
#> 4 KIRTAEREQIKTLNNKFASFIDKVRFLEQQN;NVKAQEREQIKTLNNKFASFIDKVRFLEQQN;NVKSQEREQIKTLNNKFASFIDKVRFLEQQN;QVKAQEREQIKTLNNKFASFIDKVRFLEQQN;RVKKDEKEQIKTLNNKFASFIDKVRFLEQQN;RVKTQEREQIKTLNNKFASFIDKVRFLEQQN;RVRAEEREQIKTLNNKFASFIDKVRFLEQQN;RVRQEESEQIKTLNNKFASFIDKVRFLEQQN;RVRTEEREQIKTLNNKFASFIDKVRFLEQQN;RVRTQEREQIKTLNNKFASFIDKVRFLEQQN
#> 5                                                                                                                                                                                                                                                                                                 _________MLPASPKMEHKAVVFLLLLFLK
#> 6                                                                                                                                                                                                                                                                                                                                
#>                                                                              Modification.window
#> 1                         X;X;X;X;X;X;X;X;X;X;X;X;X;X;X;Acetyl (K);X;X;X;X;X;X;X;X;X;X;X;X;X;X;X
#> 2                         X;X;X;X;X;X;X;X;X;X;X;X;X;X;X;Acetyl (K);X;X;X;X;X;X;X;X;X;X;X;X;X;X;X
#> 3             X;X;X;X;X;X;X;X;X;X;X;X;X;X;X;Acetyl (K);X;X;X;X;X;X;X;X;Oxidation (M);X;X;X;X;X;X
#> 4                         X;X;X;X;X;X;X;X;X;X;X;X;X;X;X;Acetyl (K);X;X;X;X;X;X;X;X;X;X;X;X;X;X;X
#> 5 X;X;X;X;X;X;X;X;X;Oxidation (M);X;X;X;X;X;Acetyl (K);Oxidation (M);X;X;X;X;X;X;X;X;X;X;X;X;X;X
#> 6                         X;X;X;X;X;X;X;X;X;X;X;X;X;X;X;Acetyl (K);X;X;X;X;X;X;X;X;X;X;X;X;X;X;X
#>           Peptide.window.coverage Acetyl..K..Probabilities
#> 1 XXXXXXXXXXXXPPPPPPPPPPPXXXXXXXX           HIQK(1)EDVPSER
#> 2 XXXXXXXXPPPPPPPPPPPPPPPXXXXXXXX       FALPQYLK(1)TVYQHQK
#> 3 XXXXXXXXXPPPPPPPPPPPPPPXXXXXXXX        VLPVPQK(1)AVPYPQR
#> 4 XXXXXXXXXXXPPPPPPPPPPPPPPXXXXXX        TLNNK(1)FASFIDKVR
#> 5 XXXXXXXXXPPPPPPPPPPPXXXXXXXXXXX           MLPASPK(1)MEHK
#> 6 XXXXXXPPPPPPPPPPPPPPXXXXXXXXXXX        RSALCEAATK(1)TAER
#>     Acetyl..K..Score.diffs Position.in.peptide Charge Mass.error..ppm.
#> 1          HIQK(76)EDVPSER                   4      2          1.18680
#> 2 FALPQYLK(83)TVYQHQK(-83)                   8      3          0.45024
#> 3       VLPVPQK(41)AVPYPQR                   7      3         -1.36900
#> 4  TLNNK(48)FASFIDK(-48)VR                   5      3          0.21323
#> 5     MLPASPK(61)MEHK(-61)                   7      2         -1.69530
#> 6      RSALCEAATK(100)TAER                  10      2         -3.42040
#>   Identification.type.C1_IP Identification.type.C1_prot
#> 1                                                      
#> 2                                                      
#> 3                                                      
#> 4                                                      
#> 5                                              By MS/MS
#> 6                                                      
#>   Identification.type.C2_IP Identification.type.C2_prot
#> 1                                                      
#> 2                                                      
#> 3                                                      
#> 4                                                      
#> 5                                              By MS/MS
#> 6                                              By MS/MS
#>   Identification.type.C3_IP Identification.type.C3_prot
#> 1                                                      
#> 2                                                      
#> 3                                                      
#> 4                                                      
#> 5                                              By MS/MS
#> 6                                                      
#>   Identification.type.C4_IP Identification.type.C4_prot
#> 1                                                      
#> 2                                                      
#> 3                                                      
#> 4                                                      
#> 5                                              By MS/MS
#> 6                                                      
#>   Identification.type.C5_IP Identification.type.C5_prot
#> 1                                                      
#> 2                                                      
#> 3                                                      
#> 4                                                      
#> 5                                              By MS/MS
#> 6                                                      
#>   Identification.type.W1_IP Identification.type.W1_prot
#> 1                                                      
#> 2                                                      
#> 3                                                      
#> 4                                                      
#> 5                                              By MS/MS
#> 6                                                      
#>   Identification.type.W2_IP Identification.type.W2_prot
#> 1                                                      
#> 2                                                      
#> 3                  By MS/MS                 By matching
#> 4                  By MS/MS                            
#> 5                                              By MS/MS
#> 6                                              By MS/MS
#>   Identification.type.W3_IP Identification.type.W3_prot
#> 1                                                      
#> 2                                                      
#> 3                                                      
#> 4                                                      
#> 5                                              By MS/MS
#> 6                                              By MS/MS
#>   Identification.type.W4_IP Identification.type.W4_prot
#> 1                  By MS/MS                            
#> 2                  By MS/MS                            
#> 3                  By MS/MS                            
#> 4                                                      
#> 5                                              By MS/MS
#> 6                                              By MS/MS
#>   Identification.type.W5_IP Identification.type.W5_prot  Intensity
#> 1                                                          5614700
#> 2                                                         39161000
#> 3                                                        126430000
#> 4                                                          5683200
#> 5                                              By MS/MS 1100300000
#> 6                                              By MS/MS 4191800000
#>   Intensity___1 Intensity___2 Intensity___3 Ratio.mod.base Intensity.C1_IP
#> 1       5614700             0             0     0.00070358               0
#> 2      39161000             0             0            NaN               0
#> 3     126430000             0             0            NaN               0
#> 4       5683200             0             0            NaN               0
#> 5    1100300000             0             0            NaN               0
#> 6    4191800000             0             0            NaN               0
#>   Intensity.C1_prot Intensity.C2_IP Intensity.C2_prot Intensity.C3_IP
#> 1                 0               0                 0               0
#> 2                 0               0                 0               0
#> 3                 0               0                 0               0
#> 4                 0               0                 0               0
#> 5         104990000               0         164710000               0
#> 6                 0               0         656870000               0
#>   Intensity.C3_prot Intensity.C4_IP Intensity.C4_prot Intensity.C5_IP
#> 1                 0               0                 0               0
#> 2                 0               0                 0               0
#> 3                 0               0                 0               0
#> 4                 0               0                 0               0
#> 5          80563000               0         143080000               0
#> 6                 0               0                 0               0
#>   Intensity.C5_prot Intensity.W1_IP Intensity.W1_prot Intensity.W2_IP
#> 1                 0               0                 0               0
#> 2                 0               0                 0               0
#> 3                 0               0                 0         3821800
#> 4                 0               0                 0         5683200
#> 5         108930000               0         123970000               0
#> 6                 0               0                 0               0
#>   Intensity.W2_prot Intensity.W3_IP Intensity.W3_prot Intensity.W4_IP
#> 1                 0               0                 0         5614700
#> 2                 0               0                 0        39161000
#> 3                 0               0                 0        42811000
#> 4                 0               0                 0               0
#> 5         109180000               0          75079000               0
#> 6         552530000               0         904500000               0
#>   Intensity.W4_prot Intensity.W5_IP Intensity.W5_prot Ratio.mod.base.C1_IP
#> 1                 0               0                 0                    0
#> 2                 0               0                 0                  NaN
#> 3                 0               0                 0                  NaN
#> 4                 0               0                 0                  NaN
#> 5         132860000               0          56923000                  NaN
#> 6         477460000               0         695980000                  NaN
#>   Ratio.mod.base.C1_prot Ratio.mod.base.C2_IP Ratio.mod.base.C2_prot
#> 1                    NaN                    0                    NaN
#> 2                    NaN                  NaN                    NaN
#> 3                    NaN                  NaN                    NaN
#> 4                    NaN                  NaN                    NaN
#> 5                    NaN                  NaN                    NaN
#> 6                    NaN                  NaN                    NaN
#>   Ratio.mod.base.C3_IP Ratio.mod.base.C3_prot Ratio.mod.base.C4_IP
#> 1                    0                    NaN                    0
#> 2                  NaN                    NaN                  NaN
#> 3                  NaN                    NaN                  NaN
#> 4                  NaN                    NaN                  NaN
#> 5                  NaN                    NaN                  NaN
#> 6                  NaN                    NaN                  NaN
#>   Ratio.mod.base.C4_prot Ratio.mod.base.C5_IP Ratio.mod.base.C5_prot
#> 1                    NaN                    0                    NaN
#> 2                    NaN                  NaN                    NaN
#> 3                    NaN                  NaN                    NaN
#> 4                    NaN                  NaN                    NaN
#> 5                    NaN                  NaN                    NaN
#> 6                    NaN                  NaN                    NaN
#>   Ratio.mod.base.W1_IP Ratio.mod.base.W1_prot Ratio.mod.base.W2_IP
#> 1                    0                    NaN                    0
#> 2                  NaN                    NaN                  NaN
#> 3                  NaN                    NaN                  NaN
#> 4                  NaN                    NaN                  NaN
#> 5                  NaN                    NaN                  NaN
#> 6                  NaN                    NaN                  NaN
#>   Ratio.mod.base.W2_prot Ratio.mod.base.W3_IP Ratio.mod.base.W3_prot
#> 1                      0                    0                    NaN
#> 2                    NaN                  NaN                    NaN
#> 3                    NaN                  NaN                    NaN
#> 4                    NaN                  NaN                    NaN
#> 5                    NaN                  NaN                    NaN
#> 6                    NaN                  NaN                    NaN
#>   Ratio.mod.base.W4_IP Ratio.mod.base.W4_prot Ratio.mod.base.W5_IP
#> 1              0.05093                      0                    0
#> 2                  NaN                    NaN                  NaN
#> 3                  NaN                    NaN                  NaN
#> 4                  NaN                    NaN                  NaN
#> 5                  NaN                    NaN                  NaN
#> 6                  NaN                    NaN                  NaN
#>   Ratio.mod.base.W5_prot Intensity.C1_IP___1 Intensity.C1_IP___2
#> 1                    NaN                   0                   0
#> 2                    NaN                   0                   0
#> 3                    NaN                   0                   0
#> 4                    NaN                   0                   0
#> 5                    NaN                   0                   0
#> 6                    NaN                   0                   0
#>   Intensity.C1_IP___3 Intensity.C1_prot___1 Intensity.C1_prot___2
#> 1                   0                     0                     0
#> 2                   0                     0                     0
#> 3                   0                     0                     0
#> 4                   0                     0                     0
#> 5                   0             104990000                     0
#> 6                   0                     0                     0
#>   Intensity.C1_prot___3 Intensity.C2_IP___1 Intensity.C2_IP___2
#> 1                     0                   0                   0
#> 2                     0                   0                   0
#> 3                     0                   0                   0
#> 4                     0                   0                   0
#> 5                     0                   0                   0
#> 6                     0                   0                   0
#>   Intensity.C2_IP___3 Intensity.C2_prot___1 Intensity.C2_prot___2
#> 1                   0                     0                     0
#> 2                   0                     0                     0
#> 3                   0                     0                     0
#> 4                   0                     0                     0
#> 5                   0             164710000                     0
#> 6                   0             656870000                     0
#>   Intensity.C2_prot___3 Intensity.C3_IP___1 Intensity.C3_IP___2
#> 1                     0                   0                   0
#> 2                     0                   0                   0
#> 3                     0                   0                   0
#> 4                     0                   0                   0
#> 5                     0                   0                   0
#> 6                     0                   0                   0
#>   Intensity.C3_IP___3 Intensity.C3_prot___1 Intensity.C3_prot___2
#> 1                   0                     0                     0
#> 2                   0                     0                     0
#> 3                   0                     0                     0
#> 4                   0                     0                     0
#> 5                   0              80563000                     0
#> 6                   0                     0                     0
#>   Intensity.C3_prot___3 Intensity.C4_IP___1 Intensity.C4_IP___2
#> 1                     0                   0                   0
#> 2                     0                   0                   0
#> 3                     0                   0                   0
#> 4                     0                   0                   0
#> 5                     0                   0                   0
#> 6                     0                   0                   0
#>   Intensity.C4_IP___3 Intensity.C4_prot___1 Intensity.C4_prot___2
#> 1                   0                     0                     0
#> 2                   0                     0                     0
#> 3                   0                     0                     0
#> 4                   0                     0                     0
#> 5                   0             143080000                     0
#> 6                   0                     0                     0
#>   Intensity.C4_prot___3 Intensity.C5_IP___1 Intensity.C5_IP___2
#> 1                     0                   0                   0
#> 2                     0                   0                   0
#> 3                     0                   0                   0
#> 4                     0                   0                   0
#> 5                     0                   0                   0
#> 6                     0                   0                   0
#>   Intensity.C5_IP___3 Intensity.C5_prot___1 Intensity.C5_prot___2
#> 1                   0                     0                     0
#> 2                   0                     0                     0
#> 3                   0                     0                     0
#> 4                   0                     0                     0
#> 5                   0             108930000                     0
#> 6                   0                     0                     0
#>   Intensity.C5_prot___3 Intensity.W1_IP___1 Intensity.W1_IP___2
#> 1                     0                   0                   0
#> 2                     0                   0                   0
#> 3                     0                   0                   0
#> 4                     0                   0                   0
#> 5                     0                   0                   0
#> 6                     0                   0                   0
#>   Intensity.W1_IP___3 Intensity.W1_prot___1 Intensity.W1_prot___2
#> 1                   0                     0                     0
#> 2                   0                     0                     0
#> 3                   0                     0                     0
#> 4                   0                     0                     0
#> 5                   0             123970000                     0
#> 6                   0                     0                     0
#>   Intensity.W1_prot___3 Intensity.W2_IP___1 Intensity.W2_IP___2
#> 1                     0                   0                   0
#> 2                     0                   0                   0
#> 3                     0             3821800                   0
#> 4                     0             5683200                   0
#> 5                     0                   0                   0
#> 6                     0                   0                   0
#>   Intensity.W2_IP___3 Intensity.W2_prot___1 Intensity.W2_prot___2
#> 1                   0                     0                     0
#> 2                   0                     0                     0
#> 3                   0                     0                     0
#> 4                   0                     0                     0
#> 5                   0             109180000                     0
#> 6                   0             552530000                     0
#>   Intensity.W2_prot___3 Intensity.W3_IP___1 Intensity.W3_IP___2
#> 1                     0                   0                   0
#> 2                     0                   0                   0
#> 3                     0                   0                   0
#> 4                     0                   0                   0
#> 5                     0                   0                   0
#> 6                     0                   0                   0
#>   Intensity.W3_IP___3 Intensity.W3_prot___1 Intensity.W3_prot___2
#> 1                   0                     0                     0
#> 2                   0                     0                     0
#> 3                   0                     0                     0
#> 4                   0                     0                     0
#> 5                   0              75079000                     0
#> 6                   0             904500000                     0
#>   Intensity.W3_prot___3 Intensity.W4_IP___1 Intensity.W4_IP___2
#> 1                     0             5614700                   0
#> 2                     0            39161000                   0
#> 3                     0            42811000                   0
#> 4                     0                   0                   0
#> 5                     0                   0                   0
#> 6                     0                   0                   0
#>   Intensity.W4_IP___3 Intensity.W4_prot___1 Intensity.W4_prot___2
#> 1                   0                     0                     0
#> 2                   0                     0                     0
#> 3                   0                     0                     0
#> 4                   0                     0                     0
#> 5                   0             132860000                     0
#> 6                   0             477460000                     0
#>   Intensity.W4_prot___3 Intensity.W5_IP___1 Intensity.W5_IP___2
#> 1                     0                   0                   0
#> 2                     0                   0                   0
#> 3                     0                   0                   0
#> 4                     0                   0                   0
#> 5                     0                   0                   0
#> 6                     0                   0                   0
#>   Intensity.W5_IP___3 Intensity.W5_prot___1 Intensity.W5_prot___2
#> 1                   0                     0                     0
#> 2                   0                     0                     0
#> 3                   0                     0                     0
#> 4                   0                     0                     0
#> 5                   0              56923000                     0
#> 6                   0             695980000                     0
#>   Intensity.W5_prot___3 Reverse Potential.contaminant id
#> 1                     0                             +  0
#> 2                     0                             +  1
#> 3                     0                             +  2
#> 4                     0                             +  3
#> 5                     0                             +  4
#> 6                     0       +                        5
#>              Protein.group.IDs                               Positions Position
#> 1                            2                                      83       83
#> 2                            3                                     181      181
#> 3                            4                                     176      176
#> 4 7;11;12;13;18;20;22;23;24;25 173;156;101;208;188;209;152;179;177;175      188
#> 5                            9                                       7        7
#> 6                           30                                      92       92
#>   Peptide.IDs Mod..peptide.IDs
#> 1       20674            22193
#> 2 13362;62791      14365;67570
#> 3       58301            62764
#> 4       53693            57802
#> 5       37695            40619
#> 6       46050            49622
#>                                                            Evidence.IDs
#> 1                                                                337380
#> 2                                                                212099
#> 3                                           950390;950391;950392;950393
#> 4                                                                873211
#> 5 623931;623932;623933;623934;623935;623936;623937;623938;623939;623940
#> 6                             748592;748593;748594;748595;748596;748597
#>                                                 MS.MS.IDs
#> 1                                                  324893
#> 2                                                  200658
#> 3                             908034;908035;908036;908037
#> 4                                                  834036
#> 5 600973;600974;600975;600976;600977;600978;600979;600980
#> 6                                                  714993
#>   Best.localization.evidence.ID Best.localization.MS.MS.ID
#> 1                        337380                     324893
#> 2                        212099                     200658
#> 3                        950392                     908037
#> 4                        873211                     834036
#> 5                        623938                     600980
#> 6                        748592                     714993
#>   Best.localization.raw.file Best.localization.scan.number
#> 1  Ref7928_AV_IP_08_20220809                         56954
#> 2  Ref7928_AV_IP_08_20220809                        142683
#> 3  Ref7928_AV_IP_08_20220809                        119517
#> 4  Ref7928_AV_IP_04_20220809                        143425
#> 5  Ref7928_AV_WP_10_20220809                         26580
#> 6  Ref7928_AV_WP_06_20220809                         94835
#>   Best.score.evidence.ID Best.score.MS.MS.ID       Best.score.raw.file
#> 1                 337380              324893 Ref7928_AV_IP_08_20220809
#> 2                 212099              200658 Ref7928_AV_IP_08_20220809
#> 3                 950391              908036 Ref7928_AV_IP_08_20220809
#> 4                 873211              834036 Ref7928_AV_IP_04_20220809
#> 5                 623935              600977 Ref7928_AV_WP_07_20220809
#> 6                 748592              714993 Ref7928_AV_WP_06_20220809
#>   Best.score.scan.number Best.PEP.evidence.ID Best.PEP.MS.MS.ID
#> 1                  56954               337380            324893
#> 2                 142683               212099            200658
#> 3                 119674               950391            908036
#> 4                 143425               873211            834036
#> 5                  25317               623935            600977
#> 6                  94835               748592            714993
#>           Best.PEP.raw.file Best.PEP.scan.number
#> 1 Ref7928_AV_IP_08_20220809                56954
#> 2 Ref7928_AV_IP_08_20220809               142683
#> 3 Ref7928_AV_IP_08_20220809               119674
#> 4 Ref7928_AV_IP_04_20220809               143425
#> 5 Ref7928_AV_WP_07_20220809                25317
#> 6 Ref7928_AV_WP_06_20220809                94835