Utility method to return metabolite counts found in compound class categories based on an input data compound class data object from the chemicalClassSurvey function The returned counts for each class category are for both the metabolite id query list and for the larger full or user-defined population of metabolite ids. This method is used in the exported chemicalClassEnrichment function

Utility method to return metabolite counts found in compound class categories based on an input data compound class data object from the chemicalClassSurvey function The returned counts for each class category are for both the metabolite id query list and for the larger full or user-defined population of metabolite ids. This method is used in the exported chemicalClassEnrichment function

getTotalFoundInCategories(classData, inferIdMapping = FALSE)

Arguments

classData

Data object returned from a call to chemicalClassSurvey This input contains lists of chemical classes that pertain to a query list of metabolites and pertaining to metabolites in a larger metabolite population.

inferIdMapping

if FALSE, the survey only reports on class annotations made directly on the input ids. If inferIdMapping is set to TRUE, the ids are cross-referenced or mapped to related ids that contain metabolite class annotations. The default is FALSE. Note that this utility method is typcally used within chemical class enrichment and is passed the value of this parameter.

Value

a list object with two keys, 'mets' and 'pop' that each has a table of metabolite or population chemical classes and metabolite counts per class. This supports the chemicalClassEnrichment function.