deprecated functions - to be removed in RaMP 4.0

Use plotPathwayResults instead of pathwayResultsPlot

Use getChemClass instead of chemicalClassSurvey

Use runEnrichChemClass instead of chemicalClassEnrichment

Use filterEnrichResults instead of FilterFishersResults

Use writeFishersResults instead of write_FishersResults

Use runEnrichPathways instead of runCombinedFisherTest

pathwayResultsPlot(
  db = RaMP(),
  pathwaysSig,
  pval = "FDR",
  perc_analyte_overlap = 0.5,
  perc_pathway_overlap = 0.5,
  min_pathway_tocluster = 3,
  text_size = 8,
  sig_cutoff = 0.05,
  interactive = FALSE
)

chemicalClassSurvey(
  db = RaMP(),
  mets,
  background = "database",
  background_type = "database",
  includeRaMPids = FALSE,
  inferIdMapping = TRUE
)

chemicalClassEnrichment(
  db = RaMP(),
  mets,
  background = "database",
  background_type = "database",
  inferIdMapping = F
)

FilterFishersResults(fishers_df, pval_type = "fdr", pval_cutoff = 0.1)

write_FishersResults(fishResults = "none", outputfile = "none", rampid = FALSE)

runCombinedFisherTest(
  db = RaMP(),
  analytes,
  NameOrIds = "ids",
  total_genes = 20000,
  min_analyte = 2,
  MCall = F,
  alternative = "less",
  min_path_size = 5,
  max_path_size = 150,
  includeRaMPids = FALSE,
  background_type = "database",
  background = "database",
  pathway_definitions = "RaMP",
  include_smpdb = FALSE
)

Arguments

db

a RaMP databse object

pathwaysSig

output of FilterFisherResults

pval

Which p value to plot, choose from Raw, FDR or Holm-adjusted

perc_analyte_overlap

Minimum overlap for pathways to be considered similar (Default = 0.2)

perc_pathway_overlap

Minimum overlap for clusters to merge (Default = 0.2)

min_pathway_tocluster

Minimum number of 'similar' pathways required to start a cluster (medoid) (Default = 3)

text_size

Scales all text in figure (Default=16)

sig_cutoff

Aesthetic, shows pvalue cutoff for significant pathways

interactive

If TRUE, return interactive plotly object instead of ggplot object

mets

a vector of source prepended metabolite ids

background

background to be used for Fisher's tests. If parameter 'background_type="database"', this parameter is ignored (default="database"); if parameter 'background_type= "file"', then 'background' should be a file name (with directory); if 'background_type="list"', then 'background' should be a vector of RaMP IDs; if 'backgroud_type="biospecimen"' then users should specify one of the following: "Blood", "Adipose tissue", "Heart", "Urine", "Brain", "Liver", "Kidney", "Saliva", and "Feces"

background_type

type of background that is input by the user. Opions are "database" if user wants all analytes from the RaMP database to be used as background; "file", if user wnats to input a file path with a list of background analytes; "list", if user wants to input a vector of analyte IDs; "biospecimen", if user wants to specify a biospecimen type (e.g. blood, adipose tissue, etc.) and have those biospecimen-specific analytes used. For genes, only the "database" option is used.

includeRaMPids

include internal RaMP identifiers (default is "FALSE")

inferIdMapping

if FALSE, the method only reports on class annotations made directly on the input ids. If inferIdMapping is set to TRUE, the input ids are cross-referenced or mapped to other existing ids that contain metabolite class annotations. Following id cross references can expand coverage if the input type is other than HMDB ids or LIPIDMAPS ids. The default value is FALSE.

fishers_df

The data frame generated by runFisherTest

pval_type

Specifies which p-value to use as the filter threshold. Permitted values are 'pval' and 'fdr' for chemical class and pathway enrichment. Pathway enrichment also includes an optional 'holm' value for holm p-value corrections. Default is 'fdr'.

pval_cutoff

return pathways where pval_type p-values are < pval_cutoff

fishResults

a data frame returned by function runCombinedFisherTest()

outputfile

name of output file

rampid

whether or not to include rampId (default is FALSE)

analytes

a vector of analytes (genes or metabolites) that need to be searched

NameOrIds

whether input is "names" or "ids" (default is "ids", must be the same for analytes and background)

total_genes

number of genes analyzed in the experiment (e.g. background) (default is 20000, with assumption that analyte_type is "genes")

min_analyte

if the number of analytes (gene or metabolite) in a pathway is < min_analyte, do not report

MCall

T/F if true, all pathways are used for multiple comparison corrections; if false, only pathways covering user analytes will be used (default is "F")

alternative

alternative hypothesis test passed on to fisher.test(). Options are two.sided, greater, or less (default is "less")

min_path_size

the minimum number of pathway members (genes and metabolites) to include the pathway in the output (default = 5)

max_path_size

the maximum number of pathway memnbers (genes and metaboltes) to include the pathway in the output (default = 150)

pathway_definitions

If "RaMP" (default), use pathway definitions within RaMP-DB. Else, supply path to gmx file containing custom pathway definitions. GMX files are a tab-separated format that contain one analyte set per column, with the name of the set in the first row, and constituent analytes in subsequent rows. Please supply a .xls or .xlsx file. If supplying pathway definitions for genes and metabolites, ensure that metabolite definitions are on tab 1, and gene definitions are on tab2.

include_smpdb

Include pathways from smpdb/hmdb in analysis. Excluded by default since definitions are highly redundant