Do fisher test for only one pathway from search result clicked on highchart

Do fisher test for only one pathway from search result clicked on highchart

runFisherTest(
  db = RaMP(),
  analytes,
  total_genes = 20000,
  NameOrIds = "ids",
  analyte_type = "metabolites",
  MCall = F,
  alternative = "less",
  min_path_size = 5,
  max_path_size = 150,
  background_type = "database",
  background = "database",
  pathway_definitions = "RaMP"
)

Arguments

db

a RaMP databse object

analytes

a vector of analytes (genes or metabolites) that need to be searched

total_genes

number of genes analyzed in the experiment (e.g. background) (default is 20000, with assumption that analyte_type is "genes")

NameOrIds

whether input is "names" or "ids" (default is "ids", must be the same for analytes and background)

analyte_type

"metabolites" or "genes" (default is "metabolites")

MCall

T/F if true, all pathways are used for multiple comparison corrections; if false, only pathways covering user analytes will be used (default is "F")

alternative

alternative hypothesis test passed on to fisher.test(). Options are two.sided, greater, or less (default is "less")

min_path_size

the minimum number of pathway members (genes and metabolites) to include the pathway in the output (default = 5)

max_path_size

the maximum number of pathway memnbers (genes and metaboltes) to include the pathway in the output (default = 150)

background_type

type of background that is input by the user. Opions are "database" if user wants all analytes from the RaMP database will be used; "file", if user wants to input a file with a list of background analytes; "list", if user wants to input a vector of analyte IDs; "biospecimen", if user wants to specify a biospecimen type (e.g. blood, adipose tissue, etc.) and have those biospecimen-specific analytes used. For genes, only the "database" option is used.

background

background to be used for Fisher's tests. If parameter 'background_type="database"', this parameter is ignored (default="database"); if parameter 'background_type= "file"', then 'background' should be a file name (with directory); if 'background_type="list"', then 'background' should be a vector of RaMP IDs; if 'backgroud_type="biospecimen"' then users should specify one of the following: "Blood", "Adipose tissue", "Heart", "Urine", "Brain", "Liver", "Kidney", "Saliva", and "Feces"

pathway_definitions

If "RaMP" (default), use pathway definitions within RaMP-DB. Else, supply path to gmx file containing custom pathway definitions. GMX files are a tab-separated format that contain one analyte set per column, with the name of the set in the first row, and constituent analytes in subsequent rows

Value

a dataframe with columns containing pathway ID, fisher's p value, user analytes in pathway, and total analytes in pathway