File containing sample information

.xlsx file containing all the required information about the all samples that will be used for the analysis. The file contains details of both input proteome and modified proteome data

data(sampleTable)

Format

An object of class "data.frame"

label

Mandatory column. Exact column names of LFQ (or iBAQ) and Intensity columns from MaxQuant output files

sample_name

Contains sample name

condition

Contains complete details of different conditions

replicate

Mandatory column. Replicate information of samples from different conditions

batch

Mandatory column. Batch information of samples from different conditions. Especially important for batch correction (if samples were prepared or analyzed in different batches)

fraction

Mandatory column. Contains information if the label corresponds to "Proteome" or "Enriched"

genotype

Optional column. Genotype information if required for PCA (Additional columns can be added, for example, time)

References

This data set was manually created RAmP package

Examples

data(sampleTable)
head(sampleTable)
#> # A tibble: 6 × 7
#>   label              sample_name condition replicate batch fraction genotype
#>   <chr>              <chr>       <chr>         <dbl> <dbl> <chr>    <chr>   
#> 1 Intensity.chm_IP_1 chm.RNAi.2  chm               1     1 Enriched chm     
#> 2 Intensity.chm_IP_2 chm.RNAi.2  chm               2     1 Enriched chm     
#> 3 Intensity.chm_IP_3 chm.RNAi.3  chm               3     2 Enriched chm     
#> 4 Intensity.chm_IP_4 chm.RNAi.4  chm               4     2 Enriched chm     
#> 5 Intensity.GST_IP_1 GST.RNAi.2  GST               1     1 Enriched GST     
#> 6 Intensity.GST_IP_2 GST.RNAi.2  GST               2     1 Enriched GST